+++ src/kalman_filter.cpp
@@ -48,14 +48,14 @@ void KalmanFilter::UpdateEKF(const VectorXd &z) {
    */
   DEBUGOUT("Radar update.");
 
-  float px = x_(0);
-  float py = x_(1);
-  float vx = x_(2);
-  float vy = x_(3);
+  double px = x_(0);
+  double py = x_(1);
+  double vx = x_(2);
+  double vy = x_(3);
 
-  float rho = sqrt(px*px + py*py);
-  float theta = atan2(py, px);
-  float rho_dot = (px*vx + py*vy) / rho;
+  double rho = sqrt(px*px + py*py);
+  double theta = atan2(py, px);
+  double rho_dot = (px*vx + py*vy) / rho;
   VectorXd h = VectorXd(3);
   h << rho, theta, rho_dot;
   VectorXd y = z - h;
